libbus1: add public bus1 macro header
This adds a new header called bus1-public.h. As it turns out, we want
some macros like B1_ALIGN() and friends to be used in our other headers.
As we really want to reduce inter-header dependencies, as as we want to
not pollute the public namespace too much, this ass bus1-public.h.

The only purpose of this header is to collect stuff we need in other
public headers. It is not limited to a fixed subsystem, but rather
collects stuff from all over the place, in case we want it in a public
header.

Signed-off-by: David Herrmann <dh.herrmann@gmail.com>