macro: document that ISO-C11 headers are API
The c-macro.h header includes a basic set of ISO-C11 headers, which we
consider API. That is, if you include c-macro.h, those C11 headers are
guaranteed to be included as well. This follows the ISO specification of
other C11 headers, which also often imply another set of headers to be
included.

Signed-off-by: David Herrmann <dh.herrmann@gmail.com>