macro: fix c_clz() with 64bit literals
If you run c_clz(), and as such _Generic() with string-literals and
evaluate it at *runtime*, any _Generic() selection that is not hit is
still compiled and might produce dead-code. This is fine and is omitted
with -O by gcc, but it still can produce compile-time warnings and errors
if you pass literals. In those cases, we get implicit truncations, which
is really bad.

Add explicit casts, which are no-ops but protect us from errors in those
dead-code paths.

Signed-off-by: David Herrmann <dh.herrmann@gmail.com>