b1-peer: always fully initialize B1Handle objects
Make sure we always initialize *all* fields of a B1Handle object.
Otherwise, we end up with stale pointers and get random behavior.

Even if a caller must adjust some fields anyway, we should always make
sure to initialize them to something proper, so an error handler can deal
with it.

While at it, make sure to keep the usual style with _c_cleanup_() helpers
on the constructed object.

Signed-off-by: David Herrmann <dh.herrmann@gmail.com>