b1-peer: interface - enforce that interfaces are immutable
For all intents and purposes interfaces should be considered
immutable once created. Moreover, the set of interfaces a given
node implements does not change at run-time.

Enforce this by refusing to add new members to an interface
that has already been implemented, and refuse to add new interfaces
to a node once it has received messages.

Signed-off-by: Tom Gundersen <teg@jklm.no>