b1-peer: message - add support for creating seed messages
Conceptually, the header of a seed message contains a map from
interface names to nodes owned by the peer. In practice, the
map is from strings to offsets into the array of attached handles.
We guarantee that the corresponding handles corresponds to nodes
owned by the peer. We do not verify the strings, but this will
be done when receiving the message.

Signed-off-by: Tom Gundersen <teg@jklm.no>