b1-peer: create root-nodes map
The map contains named root-nodes that have not yet been implemented
by the peer. These will have been recevied from a seed message, and
once their expected interface is implemented, they are removed from
the root-nodes map.

This is not yet hooked up, and for now we just use it to return
the correct error message in case a method call is made on an interface
that should have been, but has not been, implemented.

Signed-off-by: Tom Gundersen <teg@jklm.no>