b1-peer: handle/node - rework creation and linking
Require the ids to be set before calling _link(), and take the ids
as arguments when calling _new(). Moreover, split the owner handle
creation from the node creation in node_new_internal().

Also explicitly require the owner/holder to always be set for all
handles/nodes. This was already the case, so delete the redundant
fall-back paths.

This should have no functional change.

Signed-off-by: Tom Gundersen <teg@jklm.no>