b1-peer: dispatch - add seed support
A seed message is dispatched by simply replacing the root nodes
of the peer with the root nodes of the message. Note that this
does not affect root nodes that have already been implemented,
and it means the peer gets unique ownership of the message's
root nodes, so dispatching the same message a second time would
be a noop.

Signed-off-by: Tom Gundersen <teg@jklm.no>