message: reply - don't allow reply to reply
Our typesysetm does not deal with this, so we would either need to come
up with a more powerful type system or we simply don't allow it. For
now we did not really have a compelling usecase for 'conversations', but
this is something we could add in the future if the need arises.

Signed-off-by: Tom Gundersen <teg@jklm.no>