tests: activator - start adding an activator sample
This adds the Manager and Component objects.

The manager contains the root peer that all other peers are cloned
from and maps of Components and Dependencies (to be implemented).

Each Component is named and indexd by the name, but this is just
for convenience, and semantically they should be considered unnamed.
A component encapsulates a peer, which is the cloned child of the
root peer, and which will be passed on to the forked proces when
the component is spawned.

A communication channel is created between the root peer and each
child, represented by a node in the parent and a handle in the child.
Each such node/handle pair is stored in the corresponding component
object. The intention is that the node will implement an interface,
which the child will communicate with using the handle in order to
bootstrap itself.

Signed-off-by: Tom Gundersen <teg@jklm.no>