tests: activator - add dependency tracking
Components do not depend on eachother directly, but on eachother's
named root nodes. Or in other words, on the root interfaces they
each implement.

We have a global map in the manager from interface names to handles,
and we allow handles to root nodes to be pushed from components to
the manager, and hanldes representing dependencies to fetched by
the component from the manager.

Note that all this setup should happen in the activator process, and
the privileged communication channel used should be destroyed before
the component is actually spawned.

Currently, we implement this in terms of a full-fledged interface,
this is clearly overkill as messages could be processed directly
as they are inside the same process, rather than handing them off
to be dispatched. However, this seemed clearer for me to get my
head around, and we can short-circuit it once we have it all down.

Signed-off-by: Tom Gundersen <teg@jklm.no>