tests: activator - initialize dependencies
Fetch the dependencies from the activator and store them together with
the root nodes as the component's seed.

The only contract between the activator and the real component is the
contents of the seed. Everything not in the seed will not survive
fork. In particular, the privileged management channel we have been
using to set everything up will be destroyed.

Signed-off-by: Tom Gundersen <teg@jklm.no>