b1-peer: match - add B1Match objects
The intended usage is that a match object contains all the reply
handles that would match some sort of filter.

When a message should be multicast, it should be compared to the
filters associtade with each match object, and it should then be
sent to all the B1Match objects whose filter match.

Signed-off-by: Tom Gundersen <teg@jklm.no>