b1-peer: message - distinguish between notification types
The kernel now distinguishes between RELEASE and DESTROY notifications,
reflect that in userspace. The former is sent to the owner of a node
once all handles to it has been released, the latter is sent to the
holder of a handle once a node has been destroyed by its owner.

Signed-off-by: Tom Gundersen <teg@jklm.no>