handle: rework refcounting
We now distinguish between references to the userspace object, and
references to the kernel object.

We pin exactly one reference in the kernel if and only if the number
of references to the kernel object is greater than zero.

We may have a userspace ref, but without a kernel ref, in case of
lazy initialization of owner nodes.

Whenever a handle is received from the kernel, b1_handle_acquire()
must be called to either allocate an object or grab a reference
to an existing one. This function is repsonsbile for maintaining
exactly one reference in the kernel.

Signed-off-by: Tom Gundersen <teg@jklm.no>